chore(popup-menu-web): split PopupMenu and create Menu component
